+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.4.0
+**BREAKING CHANGE: REQUIRES CONFIG RESET!**
+- Renamed Projectile/Explosion damage attributes (You'll see some Unknown Attribute log messages, these will only happen once per entity and can be ignored)
+- Changed projectile/explosion damage attributes to be multipliers
+- Fix Ranged/Explosion damage bonuses not scaling properly with level
+- Fix damage being modified even if it should have been blocked
+
 ## 0.3.1
 - Fix incorrect id's for projectile/explosion damage bonus